implement switchmastertosub
